requirements.txt
streamlit==1.29.0
google-generativeai==0.3.2
Pillow==10.1.0
python-magic==0.4.27
httpx==0.25.2
